+++ lib/document_picker.dart
@@ -1,3 +1,4 @@
 library document_picker;
 
 export 'source/image_selector.dart';
+export 'source/profile_picture.dart';