+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.3] - June 14, 2021
+
+* Added new view for profile picture selection
+* Added image cropper to crop image
+
 ## [0.0.2] - June 10, 2021
 
 * Add null safety support